refactor testing script
Refactor testing suite to enhance code quality, discovered bug in lexer. the lexer fails to tokenize certain files, so i am currently fixing it. this commit contains a start of me fixing it. i also added a  macro in the diagnostics header. Now, the testing suite will automatically create a log file in the event that a test fails, containing that test's output. Also, the check to ensure that a test passes is more robust. tests must return an output including the string '[PASS]' to be registered as passed. in the future, I need to look at the output code (return 0 for pass, otherwise fail). I also slightly changed the header in the readme, and updated two links in the readme so they point at the right place.